Previous owner is now an object

+++ CompulsoryCrashCourse2019/PetPrinter.cs
@@ -73,28 +73,15 @@ namespace PetShop.Console2019
                     case 13:
                         DeleteOwner();
                         break;
-                    default:
+                    case 14:
                         ExitMenu();
+                        break;
+                    default:
                         break;
                 }
 
                 Console.ReadLine();
             }
-        }
-
-        
-
-        public void PrintPets()
-        {
-           Console.WriteLine("Printing all pets!\n");
-
-
-           foreach (var pet in service.GetPets())
-           {
-               Console.WriteLine($"Name: {pet.Name}, Type: {pet.PetType}, Birthday: {pet.BirthDate}, Color: {pet.Color}, " +
-                                 $"Previous owner: {pet.PreviousOwner} Price: {pet.Price}, Sold date: {pet.SoldDate}\n");
-
-           }
         }
 
         public void MenuChoices()
@@ -118,6 +105,19 @@ namespace PetShop.Console2019
             Console.WriteLine("\nChoose a menu item!");
         }
 
+        public void PrintPets()
+        {
+            Console.WriteLine("Printing all pets!\n");
+
+
+            foreach (var pet in service.GetPets())
+            {
+                Console.WriteLine($"Id: {pet.Id}, Name: {pet.Name}, Type: {pet.PetType}, Birthday: {pet.BirthDate}, Color: {pet.Color}, " +
+                                  $"Previous owner: {pet.PreviousOwner.FirstName} {pet.PreviousOwner.LastName} ,Price: {pet.Price}, Sold date: {pet.SoldDate}\n");
+
+            }
+        }
+
         public void ReadPet()
         {
             var id = PrintPetById();
@@ -128,7 +128,7 @@ namespace PetShop.Console2019
                 if (!service.ValidateId(petToRead.Id, id))
                 {
                     Console.WriteLine($"The pet to read: Id {petToRead.Id} name: {petToRead.Name} Type: {petToRead.PetType}, Birthday: {petToRead.BirthDate}, Color: {petToRead.Color}, " +
-                                      $"Previous owner: {petToRead.PreviousOwner} Price: {petToRead.Price}, Sold date: {petToRead.SoldDate}");
+                                      $"Previous owner: {petToRead.PreviousOwner.FirstName} {petToRead.PreviousOwner.LastName} Price: {petToRead.Price}, Sold date: {petToRead.SoldDate}");
                 }
             }
             else
@@ -139,6 +139,7 @@ namespace PetShop.Console2019
 
         public void CreatePet()
         {
+
             Console.WriteLine("Enter the name of the pet: ");
             string name = Console.ReadLine();
 
@@ -155,7 +156,7 @@ namespace PetShop.Console2019
             string color = Console.ReadLine();
 
             Console.WriteLine("Enter the name of the previous owner: ");
-            string prevOwner = Console.ReadLine();
+            Owner prevOwner = NewOwner();
 
             Console.WriteLine("Enter the price of the pet");
             double price = Convert.ToDouble(Console.ReadLine());
@@ -167,7 +168,17 @@ namespace PetShop.Console2019
         {
             var id = PrintPetById();
             var petFound = service.FindPetById(id);
-            service.Delete(petFound);
+            
+            if (petFound != null)
+            {
+                service.Delete(petFound);
+                Console.WriteLine("Pet deleted!");
+            }
+            else
+            {
+                Console.WriteLine("Not a valid id!");
+            }
+            
         }
 
         public void UpdatePet()
@@ -175,30 +186,36 @@ namespace PetShop.Console2019
             var id = PrintPetById();
             var petToUpdate = service.FindPetById(id);
 
-            Console.WriteLine($"The pet to edit: Id {petToUpdate.Id} name: {petToUpdate.Name} Type: {petToUpdate.PetType}, Birthday: {petToUpdate.BirthDate}, Color: {petToUpdate.Color}, " +
-                              $"Previous owner: {petToUpdate.PreviousOwner} Price: {petToUpdate.Price}, Sold date: {petToUpdate.SoldDate}\n");
+            if (petToUpdate != null)
+            {
+                Console.WriteLine($"The pet to edit: Id {petToUpdate.Id} name: {petToUpdate.Name} Type: {petToUpdate.PetType}, Birthday: {petToUpdate.BirthDate}, Color: {petToUpdate.Color}, " +
+                                  $"Previous owner: {petToUpdate.PreviousOwner.FirstName} {petToUpdate.PreviousOwner.LastName} ,Price: {petToUpdate.Price}, Sold date: {petToUpdate.SoldDate}\n");
 
-            Console.WriteLine("Enter the name of the pet: ");
-            petToUpdate.Name = Console.ReadLine();
+                Console.WriteLine("Enter the name of the pet: ");
+                petToUpdate.Name = Console.ReadLine();
 
-            Console.WriteLine("Enter the type of pet: ");
-            petToUpdate.PetType = Console.ReadLine();
+                Console.WriteLine("Enter the type of pet: ");
+                petToUpdate.PetType = Console.ReadLine();
 
-            Console.WriteLine("Enter the birthdate of the pet: ");
-            petToUpdate.BirthDate = Convert.ToDateTime(Console.ReadLine());
+                Console.WriteLine("Enter the birthdate of the pet: ");
+                petToUpdate.BirthDate = Convert.ToDateTime(Console.ReadLine());
 
-            Console.WriteLine("Enter the date when the pet was sold: ");
-            petToUpdate.SoldDate = Convert.ToDateTime(Console.ReadLine());
+                Console.WriteLine("Enter the date when the pet was sold: ");
+                petToUpdate.SoldDate = Convert.ToDateTime(Console.ReadLine());
 
-            Console.WriteLine("Enter the color of the pet: ");
-            petToUpdate.Color = Console.ReadLine();
+                Console.WriteLine("Enter the color of the pet: ");
+                petToUpdate.Color = Console.ReadLine();
 
-            Console.WriteLine("Enter the name of the previous owner: ");
-            petToUpdate.PreviousOwner = Console.ReadLine();
+                Console.WriteLine("Enter the name of the previous owner: ");
+                petToUpdate.PreviousOwner = NewOwner();
 
-            Console.WriteLine("Enter the price of the pet");
-            petToUpdate.Price = Convert.ToDouble(Console.ReadLine());
-
+                Console.WriteLine("Enter the price of the pet");
+                petToUpdate.Price = Convert.ToDouble(Console.ReadLine());
+            }
+            else
+            {
+                Console.WriteLine("Not a valid id!");
+            }
         }
 
         public void GetPetByName()
@@ -210,7 +227,7 @@ namespace PetShop.Console2019
             if (petToGet != null)
             {
                 Console.WriteLine($"The pet you are looking for: Id {petToGet.Id} name: {petToGet.Name} Type: {petToGet.PetType}, Birthday: {petToGet.BirthDate}, Color: {petToGet.Color}, " +
-                                  $"Previous owner: {petToGet.PreviousOwner} Price: {petToGet.Price}, Sold date: {petToGet.SoldDate}\n");
+                                  $"Previous owner: {petToGet.PreviousOwner.FirstName} {petToGet.PreviousOwner.LastName}, Price: {petToGet.Price}, Sold date: {petToGet.SoldDate}\n");
 
             }
             else
@@ -229,12 +246,13 @@ namespace PetShop.Console2019
                 if (petType != null && pet.PetType.ToLower().Equals(petType.ToLower()))
                 {
                     Console.WriteLine($"The type of pet you are looking for: Id {pet.Id} name: {pet.Name} Type: {pet.PetType}, Birthday: {pet.BirthDate}, Color: {pet.Color}, " +
-                                      $"Previous owner: {pet.PreviousOwner} Price: {pet.Price}, Sold date: {pet.SoldDate}\n");
+                                      $"Previous owner: {pet.PreviousOwner.FirstName} {pet.PreviousOwner.LastName}, Price: {pet.Price}, Sold date: {pet.SoldDate}\n");
                 }
-
+                else
+                {
+                    Console.WriteLine("Not a valid type!");
+                }
             }
-            Console.WriteLine("Not a valid type!");
-
         }
 
         public void SortByPrice()
@@ -252,7 +270,7 @@ namespace PetShop.Console2019
             foreach (var pet in fivePets)
             {
                 Console.WriteLine($"The five cheapest pets are: Id {pet.Id} name: {pet.Name} Type: {pet.PetType}, Birthday: {pet.BirthDate}, Color: {pet.Color}, " +
-                                  $"Previous owner: {pet.PreviousOwner} Price: {pet.Price}, Sold date: {pet.SoldDate}\n");
+                                  $"Previous owner: {pet.PreviousOwner.FirstName} {pet.PreviousOwner.LastName}, Price: {pet.Price}, Sold date: {pet.SoldDate}\n");
             }
 
         }
@@ -283,7 +301,7 @@ namespace PetShop.Console2019
             }
         }
 
-        public void NewOwner()
+        public Owner NewOwner()
         {
             Console.WriteLine("Enter firstname: ");
             var firstName = Console.ReadLine();
@@ -300,7 +318,7 @@ namespace PetShop.Console2019
             Console.WriteLine("Enter email: ");
             var email = Console.ReadLine();
 
-            oService.AddOwner(firstName, lastName, address, phoneNr, email);
+            return oService.AddOwner(firstName, lastName, address, phoneNr, email);
         }
 
         public void UpdateOwner()
@@ -308,30 +326,46 @@ namespace PetShop.Console2019
             var id = PrintOwnerById();
             var ownerToUpdate = oService.FindOwnerById(id);
 
-            Console.WriteLine($"Firstname: {ownerToUpdate.FirstName}, Lastname: {ownerToUpdate.LastName}, Address: {ownerToUpdate.Address}, " +
-                              $"Phone number: {ownerToUpdate.PhoneNumber}, Email: {ownerToUpdate.Email}");
+            if (ownerToUpdate != null)
+            {
+                Console.WriteLine($"Firstname: {ownerToUpdate.FirstName}, Lastname: {ownerToUpdate.LastName}, Address: {ownerToUpdate.Address}, " +
+                                  $"Phone number: {ownerToUpdate.PhoneNumber}, Email: {ownerToUpdate.Email}");
 
-            Console.WriteLine("Enter firstname: ");
-            ownerToUpdate.FirstName = Console.ReadLine();
+                Console.WriteLine("Enter firstname: ");
+                ownerToUpdate.FirstName = Console.ReadLine();
 
-            Console.WriteLine("Enter lastname: ");
-            ownerToUpdate.LastName = Console.ReadLine();
+                Console.WriteLine("Enter lastname: ");
+                ownerToUpdate.LastName = Console.ReadLine();
 
-            Console.WriteLine("Enter address: ");
-            ownerToUpdate.Address = Console.ReadLine();
+                Console.WriteLine("Enter address: ");
+                ownerToUpdate.Address = Console.ReadLine();
 
-            Console.WriteLine("Enter phone number: ");
-            ownerToUpdate.PhoneNumber = Console.ReadLine();
+                Console.WriteLine("Enter phone number: ");
+                ownerToUpdate.PhoneNumber = Console.ReadLine();
 
-            Console.WriteLine("Enter email: ");
-            ownerToUpdate.Email = Console.ReadLine();
+                Console.WriteLine("Enter email: ");
+                ownerToUpdate.Email = Console.ReadLine();
+            }
+            else
+            {
+                Console.WriteLine("Not a valid id!");
+            }
         }
 
         public void DeleteOwner()
         {
             var id = PrintOwnerById();
             var ownerFound = oService.FindOwnerById(id);
-            oService.RemoveOwner(ownerFound);
+ 
+            if (ownerFound != null && !ownerFound.Equals(" "))
+            {
+                oService.RemoveOwner(ownerFound);
+                Console.WriteLine("Owner deleted!");
+            }
+            else
+            {
+                Console.WriteLine("Sorry! we could not find the id!");
+            }
         }
 
         public int PrintOwnerById()
